Rewrote the outliers rejection algorithm so besides thresholding Cook's Distance at 3. Now the program would automatically attempts to remove the point with the highest D until a R^2 greater than minimum required is reached.
git-svn-id: https://prismsvn.pnl.gov/svn/User_Applications/Crowell_Utilities/ImsInformed/trunk@19547 91f24f89-0e13-0410-88a5-a0bb1538c76e

+++ ImsInformed/Parameters/MoleculeWorkflowParameters.cs
@@ -54,6 +54,6 @@ namespace ImsInformed.Parameters
         /// <summary>
         /// Gets or sets the confidence threshold.
         /// </summary>
-        public double MinFitPoints { get; set; }
+        public int MinFitPoints { get; set; }
     }
 }
